+++ components/aws/common/common.mjs
