+++ components/aws/sources/new-emails-sent-to-ses-catch-all-domain/new-emails-sent-to-ses-catch-all-domain.mjs
@@ -1,6 +1,6 @@
 import { v4 as uuid } from "uuid";
 import base from "../common/ses.mjs";
-import { toSingleLineString } from "../common/utils.mjs";
+import { toSingleLineString } from "../../common/utils.mjs";
 import { MailParser } from "mailparser-mit";
 
 export default {
